Rename to production class names
